Revert "Implement Complete Canvas Training Module with Interactive Tactical Board"
