+++ src/components/treinos/index.ts
@@ -1,5 +1,1 @@
 export { TreinosPage } from './TreinosPage';
-export { TreinosManager } from './TreinosManager';
-export { Canvas } from './Canvas';
-export { CanvasToolbar } from './CanvasToolbar';
-export { TreinoForm } from './TreinoForm';
